+++ apm-sniffer/apm-sdk-plugin/jdbc-commons/src/main/java/org/apache/skywalking/apm/plugin/jdbc/JDBCPluginConfig.java
@@ -28,7 +28,7 @@ public class JDBCPluginConfig {
              * If set to true, the parameters of the sql (typically {@link java.sql.PreparedStatement}) would be
              * collected.
              */
-            public static boolean TRACE_SQL_PARAMETERS = false;
+            public static volatile boolean TRACE_SQL_PARAMETERS = false;
             /**
              * For the sake of performance, SkyWalking won't save the entire parameters string into the tag, but only
              * the first {@code SQL_PARAMETERS_MAX_LENGTH} characters.